pyproject.toml
[project]
name = "bit-flippers"
version = "0.1.0"
description = "A SNES-inspired turn-based RPG set in a post-apocalyptic world"
requires-python = ">=3.11"
dependencies = ["pygame-ce"]

[project.scripts]
bit-flippers = "bit_flippers.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bit_flippers"]
